refactor(main): move provider in _provider
